Use SpigotMC BuildTools by default
My (`SpraxDev`) BuildTool seems to cause some trouble when not up-to-date to SpigotMC's and depending on the build environement. I decided to use the original for now.

You'll probably get more information at https://github.com/SpraxDev/Spigot-BuildTools/releases. Maybe I'll discontinue it as it is not worth the effort.