Better Java 16 / Minecraft 1.17+ support
Using a version older than Java 16 with the original BuildTools causes the build to fail as Minecraft 1.17 requires Java 16 and Tool preparations are done without telling the tool a version to build (until now).

Additionally the test builds in this repo have been updated to use java 16 for the latest build.