Fixes concurrency bug in Spigot BuildTools related git config
If we are running multiple BuildTools instances at the same time,
they might try configuring the global git config at the same time
resulting in the processing exiting with a failure.

We now backup the current configured values and set our owns.
At the end of the Action, we restore the previous values.

That way BuildTools does not crash trying to do it itself and a nice side effect is
that Commits that might be made later in the Workflow
don't show up as made by the BuildTools.

This has been bothering me for 2 years or more... And there's a ticket for it from 2014 \o/
https://hub.spigotmc.org/jira/browse/BUILDTOOLS-26