feat(sftp-cache): add optional sftp cache to speed up cross-branch build
This allows to use an external cache for spigot build artifacts
in addition (or in place of) GitHub's actions/cache
The SFTP-Cache can live for longer and be shared more aggressively across
branches etc.

Building Spigot versions is expensive, building multiple of them even more!
I feel this a lot in my projects...